FPS v2.0.1 NeedsCap -> gLit_GUI

+++ 3dLand/GlLand/gLit_GUI/src/headers/window.h
@@ -16,7 +16,7 @@ class Window {
 
     const double maxFPS = 244.0, maxPeriod = 1.0/maxFPS;
     float deltaTime=0.f,endTime=0.f;
-    double FPSRate;
+    float FPSRate;
     int fpsEpochs=0;
 
     float uptime;
